+++ src/JsonApi/Response/FetchResponse.php
@@ -3,7 +3,7 @@ namespace WoohooLabs\Yin\JsonApi\Response;
 
 use Psr\Http\Message\ResponseInterface;
 use WoohooLabs\Yin\JsonApi\Request\RequestInterface;
-use WoohooLabs\Yin\JsonApi\Transformer\AbstractCompoundDocument;
+use WoohooLabs\Yin\JsonApi\Transformer\AbstractSuccessfulDocument;
 use WoohooLabs\Yin\JsonApi\Transformer\AbstractErrorDocument;
 
 class FetchResponse extends AbstractResponse
@@ -24,11 +24,11 @@ class FetchResponse extends AbstractResponse
      * "A server MUST respond to a successful request to fetch an individual resource or resource
      * collection with a 200 OK response."
      *
-     * @param \WoohooLabs\Yin\JsonApi\Transformer\AbstractCompoundDocument $document
+     * @param \WoohooLabs\Yin\JsonApi\Transformer\AbstractSuccessfulDocument $document
      * @param mixed $domainObject
      * @return \Psr\Http\Message\ResponseInterface
      */
-    public function ok(AbstractCompoundDocument $document, $domainObject)
+    public function ok(AbstractSuccessfulDocument $document, $domainObject)
     {
         return $this->getDocumentResourceResponse($this->request, $this->response, $document, $domainObject, 200);
     }
@@ -40,11 +40,11 @@ class FetchResponse extends AbstractResponse
      * "A server MUST respond to a successful request to fetch an individual resource or resource
      * collection with a 200 OK response."
      *
-     * @param \WoohooLabs\Yin\JsonApi\Transformer\AbstractCompoundDocument $document
+     * @param \WoohooLabs\Yin\JsonApi\Transformer\AbstractSuccessfulDocument $document
      * @param mixed $domainObject
      * @return \Psr\Http\Message\ResponseInterface
      */
-    public function okWithMeta(AbstractCompoundDocument $document, $domainObject)
+    public function okWithMeta(AbstractSuccessfulDocument $document, $domainObject)
     {
         return $this->getDocumentMetaResponse($this->response, $document, $domainObject, 200);
     }