+++ tests/JsonApi/Response/DeleteResponseTest.php
@@ -4,7 +4,7 @@ namespace WoohooLabsTest\Yin\JsonApi\Response;
 use PHPUnit_Framework_TestCase;
 use WoohooLabs\Yin\JsonApi\Request\Request;
 use WoohooLabs\Yin\JsonApi\Response\DeleteResponse;
-use WoohooLabsTest\Yin\JsonApi\Utils\StubCompoundDocument;
+use WoohooLabsTest\Yin\JsonApi\Utils\StubSuccessfulDocument;
 use Zend\Diactoros\ServerRequest;
 use Zend\Diactoros\Response;
 
@@ -13,7 +13,7 @@ class DeleteResponseTest extends PHPUnit_Framework_TestCase
     public function testOkWithMeta()
     {
         $meta = ["abc" => "def"];
-        $document = new StubCompoundDocument([], [], null, $meta);
+        $document = new StubSuccessfulDocument([], [], null, $meta);
 
         $response = $this->createResponse()->okWithMeta($document, []);
         $this->assertEquals(200, $response->getStatusCode());