+++ tests/JsonApi/Response/FetchRelationshipResponseTest.php
@@ -5,7 +5,7 @@ use PHPUnit_Framework_TestCase;
 use WoohooLabs\Yin\JsonApi\Request\Request;
 use WoohooLabs\Yin\JsonApi\Response\FetchRelationshipResponse;
 use WoohooLabs\Yin\JsonApi\Transformer\ErrorDocument;
-use WoohooLabsTest\Yin\JsonApi\Utils\StubCompoundDocument;
+use WoohooLabsTest\Yin\JsonApi\Utils\StubSuccessfulDocument;
 use Zend\Diactoros\ServerRequest;
 use Zend\Diactoros\Response;
 
@@ -13,7 +13,7 @@ class FetchRelationshipResponseTest extends PHPUnit_Framework_TestCase
 {
     public function testOk()
     {
-        $document = new StubCompoundDocument();
+        $document = new StubSuccessfulDocument();
 
         $response = $this->createResponse()->ok($document, []);
         $this->assertEquals(200, $response->getStatusCode());