+++ tests/JsonApi/Response/UpdateRelationshipResponseTest.php
@@ -5,7 +5,7 @@ use PHPUnit_Framework_TestCase;
 use WoohooLabs\Yin\JsonApi\Request\Request;
 use WoohooLabs\Yin\JsonApi\Response\UpdateRelationshipResponse;
 use WoohooLabs\Yin\JsonApi\Transformer\ErrorDocument;
-use WoohooLabsTest\Yin\JsonApi\Utils\StubCompoundDocument;
+use WoohooLabsTest\Yin\JsonApi\Utils\StubSuccessfulDocument;
 use Zend\Diactoros\ServerRequest;
 use Zend\Diactoros\Response;
 
@@ -13,7 +13,7 @@ class UpdateRelationshipResponseTest extends PHPUnit_Framework_TestCase
 {
     public function testOk()
     {
-        $document = new StubCompoundDocument();
+        $document = new StubSuccessfulDocument();
 
         $response = $this->createResponse()->ok($document, []);
         $this->assertEquals(200, $response->getStatusCode());
@@ -22,7 +22,7 @@ class UpdateRelationshipResponseTest extends PHPUnit_Framework_TestCase
     public function testOkWithMeta()
     {
         $meta = ["abc" => "def"];
-        $document = new StubCompoundDocument([], [], null, $meta);
+        $document = new StubSuccessfulDocument([], [], null, $meta);
 
         $response = $this->createResponse()->okWithMeta($document, []);
         $this->assertEquals(200, $response->getStatusCode());